move errorUtil from utils/common to utils/error
