+++ src/services/upload/thumbnailService.ts
@@ -1,5 +1,5 @@
 import { FILE_TYPE } from 'constants/file';
-import { CustomError, errorWithContext } from 'utils/common/errorUtil';
+import { CustomError, errorWithContext } from 'utils/error';
 import { logError } from 'utils/sentry';
 import { BLACK_THUMBNAIL_BASE64 } from '../../../public/images/black-thumbnail-b64';
 import FFmpegService from 'services/ffmpegService';
