+++ src/utils/common/key.ts
@@ -2,7 +2,7 @@ import { B64EncryptionResult } from 'utils/crypto';
 import CryptoWorker from 'utils/crypto';
 import { getData, LS_KEYS } from 'utils/storage/localStorage';
 import { getKey, SESSION_KEYS } from 'utils/storage/sessionStorage';
-import { CustomError } from './errorUtil';
+import { CustomError } from '../error';
 
 export const getActualKey = async () => {
     try {
